+++ examples/basic-iframe/tina/config.js
@@ -1,4 +1,4 @@
-import { defineConfig } from 'tinacms'
+import { TinaCMS, defineConfig } from 'tinacms'
 
 export default defineConfig({
   // Example of how you can override the frontend url
@@ -68,6 +68,18 @@ export default defineConfig({
   schema: {
     collections: [
       {
+        ui: {
+          filename: {
+            slugify: (values, meta) => {
+              if (meta.template.name === 'tem1') {
+                return 'tem1/' + (values?.foo || '')
+              }
+              if (meta.template.name === 'tem2') {
+                return 'tem2/' + (values?.bar || '')
+              }
+            },
+          },
+        },
         name: 'test',
         path: 'content/test',
         label: 'Test',
@@ -91,6 +103,15 @@ export default defineConfig({
         format: 'mdx',
         frontmatterFormat: 'toml',
         frontmatterDelimiters: ['+++', '+++'],
+        ui: {
+          // Example of beforeSubmit
+          beforeSubmit: async ({ values, cms, form, tinaForm }) => {
+            return {
+              ...values,
+              lastUpdated: new Date().toISOString(),
+            }
+          },
+        },
         fields: [
           {
             label: 'Title',
@@ -114,6 +135,13 @@ export default defineConfig({
               //     return 'Too Long!!!'
               //   }
               // },
+            },
+          },
+          {
+            name: 'lastUpdated',
+            type: 'datetime',
+            ui: {
+              component: 'hidden',
             },
           },
           {