+++ packages/@tinacms/app/src/App.tsx
@@ -9,6 +9,8 @@ import Playground from './Playground'
 import config from 'TINA_IMPORT'
 // @ts-expect-error
 import schemaJson from 'SCHEMA_IMPORT'
+// @ts-expect-error
+import staticMedia from 'STATIC_MEDIA_IMPORT'
 
 const RawEditor = React.lazy(() => import('./fields/rich-text'))
 
@@ -59,6 +61,7 @@ export const TinaAdminWrapper = () => {
       {...config}
       schema={schema}
       client={{ apiUrl: __API_URL__ }}
+      staticMedia={staticMedia}
       // THis will be replaced by the version of the graphql package or --garphql-version flag. It is replaced by vite at compile time
       tinaGraphQLVersion={__TINA_GRAPHQL_VERSION__}
     >