+++ packages/@tinacms/schema-tools/src/types/index.ts
@@ -606,6 +606,10 @@ export interface Config<
            * would store content in `"<my-public-folder>/uploads"`
            */
           mediaRoot: string
+          /**
+           * Indicates that media files cannot be uploaded or removed by editors
+           */
+          static?: boolean
         }
         loadCustomStore?: never
       }
@@ -745,7 +749,7 @@ type Document = {
     extension: string
   }
 }
-export interface UICollection {
+export interface UICollection<Form = any, CMS = any, TinaForm = any> {
   /**
    * Customize the way filenames are generated during content creation
    */
@@ -759,7 +763,10 @@ export interface UICollection {
      * slugify: (values) => values.title.toLowerCase().split(" ").join("-")
      * ```
      */
-    slugify?: (values: Record<string, any>) => string
+    slugify?: (
+      values: Record<string, any>,
+      meta: { collection: Collection; template: Template }
+    ) => string
     /**
      * When set to `true`, editors won't be able to modify the filename
      */
@@ -791,6 +798,30 @@ export interface UICollection {
     document: Document
     collection: Collection<true>
   }) => string | undefined
+
+  /**
+   * This function is called before a document is created or updated. It can be used to modify the values that are saved to the CMS. It can also be used to perform side effects such as sending a notification or triggering a build.
+   *
+   * @example
+   *
+   *
+   *```js
+   * beforeSubmit: async ({ values }) => {
+   *   return {
+   *     ...values,
+   *     lastUpdated: new Date().toISOString(),
+   *   };
+   * },
+   *```
+   *
+   *
+   *
+   */
+  beforeSubmit?: (arg: {
+    values: Record<string, unknown>
+    cms: CMS
+    form: TinaForm
+  }) => Promise<void | Record<string, unknown>>
 }
 
 export type DefaultItem<ReturnType> = ReturnType | (() => ReturnType)