+++ packages/tinacms/src/admin/pages/CollectionCreatePage.tsx
@@ -37,9 +37,14 @@ const createDocument = async (
   const api = new TinaAdminApi(cms)
   const { filename, ...leftover } = values
 
-  const relativePath = `${folder ? `${folder}/` : ''}${filename}.${
-    collection.format
-  }`
+  if (typeof filename !== 'string') {
+    throw new Error('Filename must be a string')
+  }
+
+  // Append the folder if it exists and the filename does not start with a slash
+  const appendFolder = folder && !filename.startsWith('/') ? `/${folder}/` : '/'
+  const relativePath = `${appendFolder}${filename}.${collection.format}`
+
   const params = api.schema.transformPayload(collection.name, {
     _collection: collection.name,
     ...(template && { _template: template.name }),
@@ -192,11 +197,17 @@ export const RenderForm = ({
       extraSubscribeValues: { active: true, submitting: true, touched: true },
       onChange: (values) => {
         if (!values?.submitting) {
+          const filename: string = values?.values?.filename
+
+          // If the filename starts with "/" then it is an absolute path and we should not append the folder name
+          const appendFolder =
+            folderName && !filename?.startsWith('/') ? `/${folderName}/` : '/'
+
           // keeps the forms relative path in sync with the filename
           form.relativePath =
             schemaCollection.path +
-            folderName +
-            `/${values?.values?.filename}.${schemaCollection.format || 'md'}`
+            appendFolder +
+            `${filename}.${schemaCollection.format || 'md'}`
         }
         if (
           slugFunction &&
@@ -204,7 +215,10 @@ export const RenderForm = ({
           !values?.submitting &&
           !values.touched?.filename
         ) {
-          const value = slugFunction(values?.values)
+          const value = slugFunction(values.values, {
+            template,
+            collection: schemaCollection,
+          })
           form.finalForm.change('filename', value)
         }
       },
@@ -246,9 +260,9 @@ export const RenderForm = ({
               return true
             }
 
-            const isValid = /^[_a-zA-Z0-9][\.\-_\/a-zA-Z0-9]*$/.test(value)
+            const isValid = /[\.\-_\/a-zA-Z0-9]*$/.test(value)
             if (value && !isValid) {
-              return 'Must begin with a-z, A-Z, 0-9, or _ and contain only a-z, A-Z, 0-9, -, _, ., or /.'
+              return 'Must contain only a-z, A-Z, 0-9, -, _, ., or /.'
             }
             // check if the filename is allowed by the collection.
             if (