+++ packages/tinacms/src/toolkit/plugin-branch-switcher/branch-switcher-legacy.tsx
@@ -104,6 +104,7 @@ export const BranchSwitcherLegacy = ({
           })
           cancelFuncs.push(cancelWaitForIndexFunc)
           waitForIndexStatusPromise
+            // @ts-ignore
             .then((indexStatus) => {
               setBranchList((previousBranchList) => {
                 // update the index status of the branch