+++ packages/tinacms/src/toolkit/react-modals/modal-provider.tsx
@@ -14,7 +14,7 @@ export const ModalProvider: React.FC = ({ children }) => {
 
   return (
     <>
-      <div id="modal-root" className="tina-tailwind" ref={setModalRef} />
+      <div id="modal-root" ref={setModalRef} />
       <ModalContainerContext.Provider
         value={{ portalNode: modalRootContainerRef }}
       >