+++ packages/@tinacms/schema-tools/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tinacms/schema-tools
 
+## 1.4.10
+
+### Patch Changes
+
+- 133e97d5b: Update the before submit types to not pass the finalForm form since it is contained in the TinaForm
+- f02b4368b: Adds a second parameter to the slugify function that passes the current collection and template.
+- 7991e097e: Add a `beforeSubmit` hook function on a collection.ui. This give users the ability to run a function before the form is submitted.
+
+  If the function returns values those values will be used will be submitted instead of the form values.
+
+  If the function returns a falsy value the original form values will be submitted.
+
+  ### Example
+
+  ```js
+  // tina/config.{ts.js}
+
+  export default defineConfig({
+    schema: {
+      collections: [
+        {
+          ui: {
+            // Example of beforeSubmit
+            beforeSubmit: async ({ values }) => {
+              return {
+                ...values,
+                lastUpdated: new Date().toISOString(),
+              }
+            },
+            //...
+          },
+          //...
+        },
+        //...
+      ],
+    },
+    //...
+  })
+  ```
+
 ## 1.4.9
 
 ### Patch Changes